+++ CHANGELOG.md
@@ -7,12 +7,19 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 ## Version 1.0.2 - Unreleased
 
 
+### Added
+* Environs as options: 
+  `XDOCTEST_VERBOSE`, `XDOCTEST_OPTIONS`, `XDOCTEST_GLOBAL_EXEC`, `XDOCTEST_REPORT`,
+  `XDOCTEST_STYLE`, and `XDOCTEST_ANALYSIS` environment variables can now be used
+  to specify configuration defaults.
+
 ### Changed
 * Added experimental hidden feature `--insert-skip-directive-above-failures`
   that can be used to modify your code such that failing doctests are marked as
   skip.
 * Disabled traceback suppression on module import errors (this might become
   configurable). 
+
 
 
 ## Version 1.0.1 - Released 2022-07-10